a little reorg :-p
