+++ train_common.cu
@@ -1,0 +1,4 @@
+
+#define TRAIN_CUDA // use cuda version of structs
+
+#include "train_common.c"
